backend/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Pillow==10.1.0
pytesseract==0.3.10
google-generativeai==0.3.2
bcrypt==4.1.2
pymongo==4.6.1
numpy==1.26.3
scikit-learn==1.4.0
python-multipart==0.0.6
Werkzeug==3.0.1
opencv-python==4.8.1.78
pdf2image==1.16.3
scikit-image==0.22.0
python-dateutil==2.8.2
gunicorn==21.2.0